+++ app/src/main/java/com/battlelancer/seriesguide/dataliberation/AutoBackupFragment.kt
@@ -17,7 +17,6 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import com.battlelancer.seriesguide.R
-import com.battlelancer.seriesguide.SgApp
 import com.battlelancer.seriesguide.databinding.FragmentAutoBackupBinding
 import com.battlelancer.seriesguide.dataliberation.DataLiberationFragment.LiberationResultEvent
 import com.battlelancer.seriesguide.dataliberation.JsonExportTask.Export
@@ -68,10 +67,10 @@ class AutoBackupFragment : Fragment() {
 
         binding.buttonAutoBackupNow.setOnClickListener {
             if (TaskManager.tryBackupTask(requireContext())) {
-                setProgressLock(true)
+                viewModel.setInProgress(true)
             }
         }
-        binding.buttonAutoBackupImport.setOnClickListener { runAutoBackupImport() }
+        binding.buttonAutoBackupImport.setOnClickListener { viewModel.runImportTask() }
 
         binding.checkBoxAutoBackupCreateCopy.isChecked =
             BackupSettings.isCreateCopyOfAutoBackup(requireContext())
@@ -115,12 +114,13 @@ class AutoBackupFragment : Fragment() {
 
         binding.groupState.visibility = View.GONE
         viewModel.updateCopiesFileNames()
-        setProgressLock(false) // Also disables import button if backup availability unknown.
 
-        // restore UI state
-        if (viewModel.isImportTaskNotCompleted) {
-            setProgressLock(true)
+        viewLifecycleOwner.lifecycleScope.launch {
+            viewModel.isInProgress.collect {
+                setProgressLock(it)
+            }
         }
+
         viewModel.availableBackupLiveData
             .observe(viewLifecycleOwner, { availableBackupTimeString: String? ->
                 val lastBackupTimeString =
@@ -157,6 +157,16 @@ class AutoBackupFragment : Fragment() {
                     }
                 }
             })
+
+        viewLifecycleOwner.lifecycleScope.launch {
+            viewModel.importSummaryState.collect { state ->
+                state.applyTo(
+                    viewModel,
+                    binding.textViewAutoBackupImportSummary,
+                    binding.scrollViewAutoBackup
+                )
+            }
+        }
     }
 
     override fun onStart() {
@@ -183,6 +193,7 @@ class AutoBackupFragment : Fragment() {
         binding = null
     }
 
+    // This is currently only posted by the export task
     @Subscribe(threadMode = ThreadMode.MAIN)
     fun onEvent(event: LiberationResultEvent) {
         event.handle(view)
@@ -193,14 +204,7 @@ class AutoBackupFragment : Fragment() {
         viewModel.updateAvailableBackupData()
         // Note: backup may remove a copy file URI
         viewModel.updateCopiesFileNames()
-        setProgressLock(false)
-    }
-
-    private fun runAutoBackupImport() {
-        setProgressLock(true)
-
-        val importTask = JsonImportTask(requireContext())
-        viewModel.importTask = SgApp.coroutineScope.launch { importTask.run() }
+        viewModel.setInProgress(false)
     }
 
     private val createShowExportFileResult =
@@ -235,6 +239,10 @@ class AutoBackupFragment : Fragment() {
         binding.buttonAutoBackupImport.isEnabled = isBackupAvailableForImport
     }
 
+    /**
+     * Also disables import button if backup availability unknown, so make sure to also call after
+     * view setup.
+     */
     private fun setProgressLock(isLocked: Boolean) {
         val binding = binding ?: return
         binding.buttonAutoBackupNow.isEnabled = !isLocked