+++ CHANGELOG.md
@@ -8,6 +8,18 @@
 Releases marked with 🧪 (or previously with the "beta" suffix) were released on
 [the preview program](https://www.seriesgui.de/help/how-to/basics/preview) only.
 
+## Version 2026.3
+
+* 🔧 Import: support importing movies by IMDB IDs (linked as external ID on TMDB), such as from the TV Time to SeriesGuide transformer.
+
+### 2026.3.0 - 2026-07-10
+
+* 🔧 Import: support importing movies with only IMDB IDs, if it is linked as an external ID on the movie's TMDB entry.
+* 🔧 Import: support new "imdb-movie" list item type, maps to added movie using "externalId" as IMDB ID on import, otherwise skips.
+* 🔧 Import: display a simple summary of imported and skipped items.
+* 🔧 Import: restore original data on failure.
+* 🔨 Export: resolve "Illegal seek" error for the DAVx5 WebDAV-backed storage provider and likely others. Thanks @mvanhorn for the tip!
+
 ## Version 2026.2
 
 * 🌟 Lists: movies can be added to lists, currently only from the details view.